requirements.txt
anthropic
fastapi
uvicorn
python-dotenv
lancedb
openai
